+++ src/Omni.AutoApi.AspNetCore/AutoApiControllerConvention.cs
@@ -1,10 +1,10 @@
-using Omni.AutoApi.Routing;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ActionConstraints;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.Extensions.Logging;
+using Omni.AutoApi.Routing;
 
 namespace Omni.AutoApi.AspNetCore
 {