+++ src/Omni.AutoApi.Client.SourceGenerator/AutoApiClientGenerator.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Text;
-using Omni.AutoApi.Routing;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
+using Omni.AutoApi.Routing;
 
 namespace Omni.AutoApi.Client.SourceGenerator
 {